backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography
email-validator
jinja2
bcrypt
python-multipart
itsdangerous
pandas
langchain
langchain_openai
umap
scikit-learn
langchain-teddynote
langchain_community
faiss_cpu
pymupdf4llm
umap_learn==0.5.7
beautifulsoup4
